pyproject.toml
[project]
name = "langgraph-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-groq>=1.1.1",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.5",
    "pillow>=12.0.0",
    "pydentic>=0.0.1.dev3",
    "python-dotenv>=1.2.1",
]
